Fix Essence Berry
This commit will address https://github.com/GTNewHorizons/GT-New-Horizons-Modpack/issues/17515

The issue is caused by IC2 since it doesn't actually account for wildcards. EnderIO (added as an optional runtime dependency) is the one who registers the itemSkull ore dict, and uses a wildcard for the metadata of the vanilla skulls.